re PR tree-optimization/44507 (vectorization ANDs array elements together incorrectly)
	PR tree-optimization/44507
	* gcc.dg/vect/pr44507.c (seeIf256ByteArrayIsConstant): Correct cast.

From-SVN: r160821

+++ gcc/testsuite/gcc.dg/vect/pr44507.c
@@ -16,7 +16,7 @@ int seeIf256ByteArrayIsConstant(
            index < 64;
            index += (int)sizeof(unsigned int))
         {
-          curVal = *((unsigned long *)(&pArray[index]));
+          curVal = *((unsigned int *)(&pArray[index]));
           orVal = orVal | curVal;
           andVal = andVal & curVal;
         }
